+++ tests/test_scores.py
@@ -18,13 +18,13 @@ async def test_scores_fetching_lxns(maimai: MaimaiClient, lxns: LXNSProvider, lx
 
     bests_fallback = await maimai.bests(lxns_player, provider=lxns)
     assert my_scores.rating == bests_fallback.rating
-    assert len(bests_fallback.scores) <= 50
+    assert len(bests_fallback.scores_b35) <= 35
 
     preview = await maimai.minfo(1231, PlayerIdentifier(friend_code=664994421382429), provider=lxns)
     assert preview is not None
     assert all(score.id == preview.song.id for score in preview.scores)
 
-    for song, diff, score in await my_scores.get_scores():
+    for song, diff, score in await my_scores.get_mapping():
         assert song.id == score.id and diff.type == score.type and diff.level_index == score.level_index
 
 
@@ -37,7 +37,7 @@ async def test_scores_fetching_divingfish(maimai: MaimaiClient, divingfish: Divi
 
     bests = await maimai.bests(divingfish_player, provider=divingfish)
     assert my_scores.rating == bests.rating
-    assert len(bests.scores) <= 50
+    assert len(bests.scores_b15) <= 15
 
     preview = await maimai.minfo(1231, PlayerIdentifier(username="turou"), provider=divingfish)
     assert preview is not None